+++ front_tributario/src/components/molecules/SimplesNacionalForm/SimplesNacionalForm.tsx
@@ -1,7 +1,7 @@
 "use client"
 import { saveSimplesNacionalForm } from "@/lib/actions";
-import { AnexoSimplesNacional, TaxType } from "@/types/types";
-import { Center, Checkbox, CheckboxGroup, Divider, FormErrorMessage, Link, Menu, MenuButton, MenuItem, MenuList, Text } from "@chakra-ui/react";
+import { AnexoSimplesNacional, IFormImput, TaxType } from "@/types/types";
+import { Center, Checkbox, CheckboxGroup, Divider, FormErrorMessage, GridItem, Link, Menu, MenuButton, MenuItem, MenuList, SimpleGrid, Text } from "@chakra-ui/react";
 import { FormHelperText } from "@chakra-ui/react";
 import { Heading } from "@chakra-ui/react";
 import { Card } from "@chakra-ui/react";
@@ -10,6 +10,10 @@ import { useEffect, useState } from "react";
 import { useFormState, useFormStatus } from 'react-dom';
 import { LuAlarmClock, LuCheckCircle2 } from "react-icons/lu";
 import { MdArrowDropDown } from "react-icons/md";
+import { TbRuler } from "react-icons/tb";
+import Select_TaxesReplaced from "./Select_TaxesReplaced/Select_TaxesReplaced";
+import Input_Value from "./Input_Value/Input_Value";
+import Select_AnnexOption from "./Select_AnnexOption/Select_AnnexOption";
 
 
 
@@ -35,6 +39,16 @@ export default function SimplesNacionalForm() {
   const [isValidValuePisCofinsReplacement, setIsValidValuePisCofinsReplacement] = useState<boolean>(true);
   const [isValidRbt12, setIsValidRbt12] = useState<boolean>(true);
   const [rbt12, setRbt12] = useState<string>("");
+  const [isIcmsSt, setIsIcmsSt] = useState(false);
+  const [isPisCofinsSt, setIsPisCofinsSt] = useState(false);
+  const [isIssSt, setIsIssSt] = useState(false);
+  const [checkedValues, setCheckedValues] = useState<TaxType[]>([]);
+
+  useEffect(() => {
+    setIsIcmsSt(checkedValues.includes("ICMS"))
+    setIsPisCofinsSt(checkedValues.includes("PIS COFINS"))
+    setIsIssSt(checkedValues.includes("ISS"))
+  }, [checkedValues])
 
   const formProps = [
     {
@@ -51,7 +65,7 @@ export default function SimplesNacionalForm() {
     },
     {
       id: "rbt12",
-      label: "Rbt12",
+      label: "RBT12",
       placeHolder: "R$ 0,00",
       helperText: "Faturamento Acumulado dos últimos 12 meses",
       validate: validateRbt12,
@@ -75,7 +89,7 @@ export default function SimplesNacionalForm() {
     },
     {
       id: "salesValueToExterior",
-      label: "Faturamento Para o Exterior",
+      label: "Faturamento Internacional",
       placeHolder: "R$ 0,00",
       helperText: "O Faturamento não pode ser Zero",
       validate: validateSalesValueToExterior,
@@ -162,304 +176,44 @@ export default function SimplesNacionalForm() {
   }
   function validateTaxesReplaced(value: string) {
     const input = Number(value)
-    if (input > 0 && input < salesValue) {
+    const sales = Number(salesValue)
+    if (input > 0 && input < sales) {
       return true
     } else { return false }
   }
   function validateIcmsReplacement(value: string) {
     const input = Number(value)
-    if (input > 0 && input < salesValue) {
+    const sales = Number(salesValue)
+    if (input > 0 && input < sales) {
       return true
     } else { return false }
   }
   function validateIssReplacement(value: string) {
     const input = Number(value)
-    if (input > 0 && input < salesValue) {
+    const sales = Number(salesValue)
+    if (input > 0 && input < sales) {
       return true
     } else { return false }
   }
   function validatePisCofinsReplacement(value: string) {
     const input = Number(value)
-    if (input > 0 && input < salesValue) {
+    const sales = Number(salesValue)
+    if (input > 0 && input < sales) {
       return true
     } else { return false }
   }
 
 
-  interface IFormImput {
-    id: string,
-    label: string,
-    placeHolder: string,
-    helperText: string,
-    validate: (value: string) => boolean, // Adicione esta linha
-    errorText: string,
-    state: string | number | TaxType[] | undefined,
-    setState: React.Dispatch<React.SetStateAction<string | number | TaxType[]>>,
-    isValid: boolean,
-    setIsValid: React.Dispatch<React.SetStateAction<boolean>>
-    children: React.ReactNode
-  }
-
-  const InputRbt12 = ({ isValid, setIsValid, state, setState, label, placeHolder, id, helperText, errorText, validate, children }: IFormImput) => {
-    const [localState, setLocalState] = useState(state);
-
-    useEffect(() => {
-      setLocalState(state);
-    }, [state]);
-
-    const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-      setLocalState(event.target.value);
-    };
-    const handleInputBlur = (event: React.ChangeEvent<HTMLInputElement>) => {
-      if (event.target.value !== '') {
-        setIsValid(validate(event.target.value));
-        setState(event.target.value);
-      }
-    };
-
-    return (
-      <Card m="10px" bg="gray.600" p="10px" >
-        <FormControl isInvalid={!isValid} >
-          <FormLabel bgGradient="linear(to-r, orange, red)" bgClip="text" >{label}</FormLabel>
-          <Input
-            placeholder={placeHolder}
-            variant="outline"
-            type="text"
-            id={id}
-            name={id}
-            isRequired
-            onBlur={handleInputBlur}
-            onChange={handleInputChange}
-            value={localState}
-          />
-          {isValid ? <FormHelperText color="gray">{helperText}</FormHelperText> : <FormHelperText color="red">{errorText}</FormHelperText>}
-        </FormControl>
-      </Card>
-    )
-  }
-
-  const FormImput = ({ isValid, setIsValid, state, setState, label, placeHolder, id, helperText, errorText, validate }: IFormImput) => {
-    const [localState, setLocalState] = useState(state);
-
-    useEffect(() => {
-      setLocalState(state);
-    }, [state]);
-
-    const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-      setLocalState(event.target.value);
-    };
-    const handleInputBlur = (event: React.ChangeEvent<HTMLInputElement>) => {
-      if (event.target.value !== '') {
-        setIsValid(validate(event.target.value));
-        setState(event.target.value);
-      }
-    };
-
-    return (
-      <Card m="10px" bg="gray.600" p="10px" >
-        <FormControl isInvalid={!isValid} >
-          <FormLabel bgGradient="linear(to-r, orange, red)" bgClip="text" >{label}</FormLabel>
-          <Input
-            placeholder={placeHolder}
-            variant="outline"
-            type="text"
-            id={id}
-            name={id}
-            isRequired
-            onBlur={handleInputBlur}
-            onChange={handleInputChange}
-            value={localState}
-          />
-          {isValid ? <FormHelperText color="gray">{helperText}</FormHelperText> : <FormHelperText color="red">{errorText}</FormHelperText>}
-        </FormControl>
-      </Card>
-    )
-  }
-
-  const FormImputRbt12 = ({ isValid, setIsValid, state, setState, label, placeHolder, id, helperText, errorText, validate }: IFormImput) => {
-    const [localState, setLocalState] = useState(state);
-
-    useEffect(() => {
-      setLocalState(state);
-    }, [state]);
-
-    const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-      setLocalState(event.target.value);
-    };
-    const handleInputBlur = (event: React.ChangeEvent<HTMLInputElement>) => {
-      if (event.target.value !== '') {
-        setIsValid(validate(event.target.value));
-        setState(event.target.value);
-      }
-    };
-
-    return (
-      <Card bg="gray.600" p="10px" >
-        <FormControl isInvalid={!isValid} >
-          <FormLabel bgGradient="linear(to-r, orange, red)" bgClip="text" fontWeight="bold" >{label}</FormLabel>
-          <Input
-            placeholder={placeHolder}
-            variant="outline"
-            type="text"
-            id={id}
-            name={id}
-            isRequired
-            onBlur={handleInputBlur}
-            onChange={handleInputChange}
-            value={localState}
-          />
-          {isValid ? <FormHelperText color="gray">{helperText}</FormHelperText> : <FormErrorMessage color="red">{errorText}</FormErrorMessage>}
-        </FormControl>
-      </Card>
-    )
-  }
-  const FormImputSalesValue = ({ isValid, setIsValid, state, setState, label, placeHolder, id, helperText, errorText, validate }: IFormImput) => {
-    const [localState, setLocalState] = useState(state);
-
-    useEffect(() => {
-      setLocalState(state);
-    }, [state]);
-
-    const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-      setLocalState(event.target.value);
-    };
-    const handleInputBlur = (event: React.ChangeEvent<HTMLInputElement>) => {
-      if (event.target.value !== '') {
-        setIsValid(validate(event.target.value));
-        setState(event.target.value);
-      }
-    };
-
-    return (
-      <Card bg="gray.600" p="10px" minW="240px" >
-        <FormControl isInvalid={!isValid} >
-          <FormLabel bgGradient="linear(to-r, orange, red)" bgClip="text" fontWeight="bold" >{label}</FormLabel>
-          <Input
-            placeholder={placeHolder}
-            variant="outline"
-            type="text"
-            id={id}
-            name={id}
-            isRequired
-            onBlur={handleInputBlur}
-            onChange={handleInputChange}
-            value={localState}
-          />
-          {isValid ? <FormHelperText color="gray">{helperText}</FormHelperText> : <FormErrorMessage color="red">{errorText}</FormErrorMessage>}
-        </FormControl>
-      </Card>
-    )
-  }
-
-  const FormSelectAnnexOption = ({ isValid, setIsValid, state, setState, label, placeHolder, id, helperText, errorText, validate }: IFormImput) => {
-    const [localState, setLocalState] = useState(state);
-    useEffect(() => {
-      setLocalState(state);
-    }, [state]);
-
-    const anexo = AnexoSimplesNacional
-    return (
-      <Card bg="gray.600" h="116px" p="10px" minW="220px" justifyContent="space-between" alignItems="self-end" >
-        <FormControl isInvalid={!isValid} id={id} >
-          <FormLabel bgGradient="linear(to-r, orange, red)" bgClip="text" fontWeight="bold" >{label}</FormLabel>
-          <Menu >
-            <MenuButton w="100%" as={Button} fontWeight="bold" variant="primary" rightIcon={<MdArrowDropDown />} >
-              {localState || label}
-            </MenuButton>
-            <MenuList bgGradient="linear(to-r, orange, red)" borderRadius="10px">
-              {Object.entries(anexo).map(([key, value]) => (
-                <MenuItem fontWeight="bold" key={key} onClick={() => { setLocalState(value); setState(value); }}>
-                  {value}
-                </MenuItem>
-              ))}
-            </MenuList>
-          </Menu>
-          <input type="hidden" name={id} value={localState} />
-          {isValid ? <FormHelperText color="gray">{helperText}</FormHelperText> : <FormHelperText color="red">{errorText}</FormHelperText>}
-        </FormControl>
-        <Link><Text fontSize="10px" color="red">Duvidas Sobre o Anexo ?</Text> </Link>
-      </Card>
-    )
-  }
 
 
-
-
-
-
-  const [isIcmsSt, setIsIcmsSt] = useState(false);
-  const [isPisCofinsSt, setIsPisCofinsSt] = useState(false);
-  const [isIssSt, setIsIssSt] = useState(false);
-  const [checkedValues, setCheckedValues] = useState<TaxType[]>([]);
-
-
-  useEffect(() => {
-    setIsIcmsSt(checkedValues.includes("ICMS"))
-    setIsPisCofinsSt(checkedValues.includes("PIS COFINS"))
-    setIsIssSt(checkedValues.includes("ISS"))
-  }, [checkedValues])
-
-
-
-  const FormSelectTaxesReplaced = ({ isValid, label, id, helperText, errorText }: IFormImput) => {
-    const tax: TaxType[] = ["ICMS", "PIS COFINS", "ISS"];
-    const handleCheckboxChange = (value: TaxType) => {
-      if (checkedValues.includes(value)) {
-        setCheckedValues(checkedValues.filter(item => item !== value));
-      } else {
-        setCheckedValues([...checkedValues, value]);
-      }
-    };
-    return (
-      <Card bg="gray.600" h="116px" p="10px" minW="220px" justifyContent="space-between" alignItems="self-end" >
-        <FormControl isInvalid={!isValid} id={id} >
-          <FormLabel bgGradient="linear(to-r, orange, red)" bgClip="text" fontWeight="bold" >{label}</FormLabel>
-          <HStack >
-            {tax.map((tax, index) => (
-              <Checkbox
-                id={id}
-                name={id}
-                color="blue"
-                value={tax}
-                key={index}
-                isChecked={checkedValues.includes(tax)}
-                onChange={() => handleCheckboxChange(tax)}
-              >
-                {tax}
-              </Checkbox>
-            ))}
-          </HStack>
-          {isValid ? <FormHelperText color="gray">{helperText}</FormHelperText> : <FormHelperText color="red">{errorText}</FormHelperText>}
-        </FormControl>
-      </Card>
-    )
-  }
 
   return (
-
     <Center mt="40px" mb="20px">
-      <HStack alignItems="start">
-        <HStack alignItems="end">
-          <VStack>
-
-            <Card bg="gray.500">
+            <Card bg="gray.500" p="5px" >
               <form action={formAction}>
-                <FormSelectAnnexOption
-                  errorText={formProps[0].errorText}
-                  helperText={formProps[0].helperText}
-                  id={formProps[0].id}
-                  isValid={formProps[0].isValid}
-                  label={formProps[0].label}
-                  placeHolder={formProps[0].placeHolder}
-                  setIsValid={formProps[0].setIsValid}
-                  setState={formProps[0].setState}
-                  state={formProps[0].state}
-                  validate={formProps[0].validate}
-                />
-                {/** 
-                <VStack justifyContent="space-between" alignItems="start">
+                <VStack >
                   <HStack w="100%" justifyContent="start" alignItems="start">
-                    <FormSelectAnnexOption
+                    <Select_AnnexOption
                       errorText={formProps[0].errorText}
                       helperText={formProps[0].helperText}
                       id={formProps[0].id}
@@ -471,7 +225,7 @@ export default function SimplesNacionalForm() {
                       state={formProps[0].state}
                       validate={formProps[0].validate}
                     />
-                    <FormImputRbt12
+                    <Input_Value
                       errorText={formProps[1].errorText}
                       helperText={formProps[1].helperText}
                       id={formProps[1].id}
@@ -484,12 +238,9 @@ export default function SimplesNacionalForm() {
                       validate={formProps[1].validate}
                     />
                   </HStack>
-                  <VStack p="5px" border="1px solid red" w="100%" justifyContent="space-between" >
-
-                    <HStack justifyContent="space-between" w="100%">
-
-
-                      <FormImputSalesValue
+                <Divider />
+                  <HStack w="100%" justifyContent="center" alignItems="center">
+                      <Input_Value
                         errorText={formProps[2].errorText}
                         helperText={formProps[2].helperText}
                         id={formProps[2].id}
@@ -501,7 +252,7 @@ export default function SimplesNacionalForm() {
                         state={formProps[2].state}
                         validate={formProps[2].validate}
                       />
-                      <FormImputSalesValue
+                      <Input_Value
                         errorText={formProps[3].errorText}
                         helperText={formProps[3].helperText}
                         id={formProps[3].id}
@@ -513,12 +264,10 @@ export default function SimplesNacionalForm() {
                         state={formProps[3].state}
                         validate={formProps[3].validate}
                       />
-                    </HStack>
-                  </VStack>
-                  <Divider />
-                  <HStack alignItems="start">
-                    <VStack>
-                      <FormSelectTaxesReplaced
+                  </HStack>
+                <Divider />
+                  <VStack w="100%" justifyContent="center" alignItems="center">
+                      <Select_TaxesReplaced
                         errorText={formProps[4].errorText}
                         helperText={formProps[4].helperText}
                         id={formProps[4].id}
@@ -529,61 +278,62 @@ export default function SimplesNacionalForm() {
                         setState={formProps[4].setState}
                         state={formProps[4].state}
                         validate={formProps[4].validate}
+                        checkedValues={checkedValues}
+                        setCheckedValues={setCheckedValues}
                       />
-                      <HStack>
-                        <FormImputSalesValue
-                          errorText={formProps[5].errorText}
-                          helperText={formProps[5].helperText}
-                          id={formProps[5].id}
-                          isValid={formProps[5].isValid}
-                          label={formProps[5].label}
-                          placeHolder={formProps[5].placeHolder}
-                          setIsValid={formProps[5].setIsValid}
-                          setState={formProps[5].setState}
-                          state={formProps[5].state}
-                          validate={formProps[5].validate}
-                        />
-                        <FormImputSalesValue
-                          errorText={formProps[6].errorText}
-                          helperText={formProps[6].helperText}
-                          id={formProps[6].id}
-                          isValid={formProps[6].isValid}
-                          label={formProps[6].label}
-                          placeHolder={formProps[6].placeHolder}
-                          setIsValid={formProps[6].setIsValid}
-                          setState={formProps[6].setState}
-                          state={formProps[6].state}
-                          validate={formProps[6].validate}
-                        />
-                        <FormImputSalesValue
-                          errorText={formProps[7].errorText}
-                          helperText={formProps[7].helperText}
-                          id={formProps[7].id}
-                          isValid={formProps[7].isValid}
-                          label={formProps[7].label}
-                          placeHolder={formProps[7].placeHolder}
-                          setIsValid={formProps[7].setIsValid}
-                          setState={formProps[7].setState}
-                          state={formProps[7].state}
-                          validate={formProps[7].validate}
-                        />
-                      </HStack>
-                    </VStack>
-                  </HStack>
+                      <SimpleGrid gap="5px" w="100%" columns={2} row={2}>
+                        <GridItem>
+                          <Input_Value
+                            errorText={formProps[5].errorText}
+                            helperText={formProps[5].helperText}
+                            id={formProps[5].id}
+                            isValid={formProps[5].isValid}
+                            label={formProps[5].label}
+                            placeHolder={formProps[5].placeHolder}
+                            setIsValid={formProps[5].setIsValid}
+                            setState={formProps[5].setState}
+                            state={formProps[5].state}
+                            validate={formProps[5].validate}
+                          /></GridItem>
+                        <GridItem>
+                          <Input_Value
+                            errorText={formProps[6].errorText}
+                            helperText={formProps[6].helperText}
+                            id={formProps[6].id}
+                            isValid={formProps[6].isValid}
+                            label={formProps[6].label}
+                            placeHolder={formProps[6].placeHolder}
+                            setIsValid={formProps[6].setIsValid}
+                            setState={formProps[6].setState}
+                            state={formProps[6].state}
+                            validate={formProps[6].validate}
+                          /></GridItem>
+                        <GridItem>
+                          <Input_Value
+                            errorText={formProps[7].errorText}
+                            helperText={formProps[7].helperText}
+                            id={formProps[7].id}
+                            isValid={formProps[7].isValid}
+                            label={formProps[7].label}
+                            placeHolder={formProps[7].placeHolder}
+                            setIsValid={formProps[7].setIsValid}
+                            setState={formProps[7].setState}
+                            state={formProps[7].state}
+                            validate={formProps[7].validate}
+                          /></GridItem>
+                      </SimpleGrid>
+                  </VStack>
+               
                 </VStack>
-                */}
-                <Card bg="gray.400" borderRadius="20px">
+                <Card mt="10px" bg="gray.400" borderRadius="20px">
                   <Card m="10px" bg="gray.600" p="10px" borderRadius="full">
-                    <Button type="submit" disabled={pending} borderRadius="full" variant={pending ? "primary" : "secondary"}  >
+                    <Button type="submit" disabled={true} borderRadius="full" variant={!pending ? "primary" : "secondary"}  >
                       {pending ? <LuAlarmClock size="35px" opacity="0.5" /> : <LuCheckCircle2 size="35px" opacity="0.5" />}
                     </Button>
                   </Card>
                 </Card>
               </form >
             </Card >
-          </VStack>
-        </HStack>
-      </HStack>
     </Center >
   );
 }
